old-tests/submission/W3C/domhtml/HTMLLegendElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - LEGEND</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-legend-element.html#dom-legend-form" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<form action="./files/getData.pl" id="form1" method="post">
<fieldset>
<legend accesskey="b" align="top">Enter Password1:</legend>
<input id="input1" name="Password1" type="password">
</fieldset>
</form>
<fieldset>
<legend accesskey="c" align="bottom">Enter Password2:</legend>
<input id="input2" name="Password2" type="password">
</fieldset>





<script>
var domtest = async_test("The form attribute returns null if control in not within the context of      form.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vform;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("legend");
      assert_equals(nodeList.length,2);
testNode = nodeList.item(1);
      vform = testNode.form;

      assert_equals(vform, null);
    
    domtest.done();
}));

</script>
<p>The form attribute returns null if control in not within the context of 
    form. 

    Retrieve the second ELEMENT and examine its form element.</p>
<div id="log"></div>
</body></html>
